requirements.txt
langchain
langchain_community
langchain-groq
wikipedia
python-dotenv
arxiv